feat: implement EncodedQuery and QsQueryEncoder for query parameter encoding
